refactor to layered structure
